+++ README.md
@@ -1,6 +1,12 @@
 # KeFetchUp!
 
-[![Build Status](https://img.shields.io/travis/KazanExpress/kefetchup/master.svg?logo=travis)](https://travis-ci.org/KazanExpress/kefetchup) [![Coverage Status](https://coveralls.io/repos/github/KazanExpress/kefetchup/badge.svg?branch=master)](https://coveralls.io/github/KazanExpress/kefetchup?branch=master) [![npm](https://img.shields.io/npm/v/kefetchup.svg?style=flat)](https://www.npmjs.com/package/kefetchup) 
-[![npm bundle size (minified)](https://img.shields.io/bundlephobia/minzip/kefetchup.svg)]() ![dependencies (minified)](https://img.shields.io/badge/dependencies-none-yellow.svg)
+[![Build Status](https://img.shields.io/travis/KazanExpress/kefetchup/master.svg?logo=travis&style=flat-square)](https://travis-ci.org/KazanExpress/kefetchup) [![Coverage status](https://img.shields.io/coveralls/github/KazanExpress/kefetchup/master.svg?style=flat-square)](https://coveralls.io/github/KazanExpress/kefetchup?branch=master) [![npm](https://img.shields.io/npm/v/kefetchup.svg?style=flat-square)](https://www.npmjs.com/package/kefetchup) 
+[![npm bundle size (minified)](https://img.shields.io/bundlephobia/minzip/kefetchup.svg?style=flat-square)]() [![dependencies (minified)](https://img.shields.io/badge/dependencies-none-yellow.svg?style=flat-square)]()
 
 > Simple fetch client API to spice up your application
+
+`npm i -S kefetchup`
+
+## What is it?
+
+This is
