+++ package.json
@@ -81,6 +81,7 @@
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "rollup-plugin-typescript2": "^0.13.0",
+    "rollup-plugin-uglify": "^6.0.0",
     "ts-jest": "^22.0.0",
     "ts-node": "^6.0.0",
     "typescript": "^3.0.0"
